# [libc++] Optimize ctype::to{lower,upper}