--- a/libcxx/lib/abi/CHANGELOG.TXT
+++ b/libcxx/lib/abi/CHANGELOG.TXT
@@ -16,6 +16,15 @@ New entries should be added directly below the "Version" header.
 Version 21.0
 ------------
 
+* [libc++] Optimize ctype::to{lower,upper}
+
+  This patch removed __classic_upper_table() and __classic_lower_table(), which were only ever accessed from the dylib.
+
+  x86_64-unknown-gnu-linux
+  ------------------------
+  Symbol removed: _ZNSt3__15ctypeIcE21__classic_lower_tableEv
+  Symbol removed: _ZNSt3__15ctypeIcE21__classic_upper_tableEv
+
 * [libc++] Instantiate hash function externally
 
   This has multiple benefits: